First pass at htslib support via xamcpp library.
xamcpp is located in (src/utils/xamcpp). The tests are mostly
working (exp~obs), but it is segfaulting at the end of the test.
I suspect the BamWriter I implemented may have a flaw. Or
possibly we are getting a double delete, and something wrong
with the sorted temp files.