Disabling pipefail when the output is explicitely troncated
"grep -m 1" stop after the first found occurence.
"samtools view" stop with an error if its output is closed.
With pipefail, the samtools error is taken into account by the lumpyexpress script that immediately exits

The error do not always occurs:
if the information looked after by grep is at the end of the BAM file and samtools is able to complete before its output is closed, then no error occurs
But we have several BAM files that triggered this error.