The call to lumpy from lumpyexpress segfaults when the -pe flags are omitted when the insert mean and stdev are not calculated in lumpyexpress. Furthermore, the lumpy galaxy wrapper fails with an error that mean & stdev are missing.
I kept the warning, but allowed mean and stdev to be output in all cases to avoid the segfaults from lumpyexpress and avoid the error & no output from the lumpy galaxy wrapper.

This is admittedly a work-around. Lumpy should probably be edited to be able to not segfault when the -pe discordant options are missing and should gracefully output no results if mean and stdev do not have valid values.

Files checked in: scripts/lumpyexpress scripts/pairend_distro.py